Generate ico & icns client-side
